app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
cryptography>=44.0.0
argon2-cffi>=23.1.0
pydantic-settings>=2.7.0
